requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.4
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-import-export==2.1.0
django-js-asset==1.2.2
django-mdeditor==0.1.18
django-modeltranslation==0.15.1
django-redis==4.12.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
et-xmlfile==1.0.1
idna==2.9
importlib-metadata==1.6.0
jdcal==1.4.1
kombu==4.6.8
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.0.0
psycopg2-binary==2.8.4
pycurl==7.43.0
PyJWT==1.7.1
pytz==2019.3
PyVimeo==1.0.11
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
requests-toolbelt==0.9.1
six==1.10.0
sqlparse==0.3.1
tablib==1.1.0
tinydb==3.15.2
tuspy==0.2.1
urllib3==1.25.8
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
